python-service/requirements.txt
flask==3.0.3
opencv-python==4.10.0.84
face-recognition==1.3.0
numpy==1.26.4
pymongo==4.8.0
python-dotenv==1.0.1
